#161 | mild refactor of some common components
